requirements.txt
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-automl==2.1.0
google-cloud-core==1.4.4
google-cloud-storage==1.33.0
google-cloud-texttospeech==2.2.0
google-cloud-vision==2.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
idna==2.10
joblib==0.17.0
libcst==0.3.15
mypy-extensions==0.4.3
numpy==1.19.4
pandas==1.1.4
proto-plus==1.11.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
